+++ vrp-pragmatic/tests/features/mod.rs
@@ -16,4 +16,5 @@ mod relations;
 mod reload;
 mod skills;
 mod timing;
+mod unassigned;
 mod work_balance;