+++ vrp-pragmatic/tests/features/unassigned/single_reason.rs
@@ -3,20 +3,17 @@ use crate::format::solution::*;
 use crate::helpers::*;
 
 #[test]
-fn can_unassign_multi_job_due_to_capacity() {
+fn can_have_empty_detail_in_empty_solution() {
     let problem = Problem {
         plan: Plan {
-            jobs: vec![create_multi_job(
-                "multi",
-                vec![((2., 0.), 1., vec![2]), ((8., 0.), 1., vec![1])],
-                vec![((6., 0.), 1., vec![3])],
+            jobs: vec![create_delivery_job_with_skills(
+                "job1",
+                (1., 0.),
+                all_of_skills(vec!["unique_skill".to_string()]),
             )],
             ..create_empty_plan()
         },
-        fleet: Fleet {
-            vehicles: vec![create_vehicle_with_capacity("my_vehicle", vec![2])],
-            profiles: create_default_matrix_profiles(),
-        },
+        fleet: Fleet { vehicles: vec![create_default_vehicle_type()], profiles: create_default_matrix_profiles() },
         ..create_empty_problem()
     };
     let matrix = create_matrix_from_problem(&problem);
@@ -29,11 +26,11 @@ fn can_unassign_multi_job_due_to_capacity() {
             statistic: Statistic::default(),
             tours: vec![],
             unassigned: Some(vec![UnassignedJob {
-                job_id: "multi".to_string(),
+                job_id: "job1".to_string(),
                 reasons: vec![UnassignedJobReason {
-                    code: "CAPACITY_CONSTRAINT".to_string(),
-                    description: "does not fit into any vehicle due to capacity".to_string(),
-                    detail: None,
+                    code: "SKILL_CONSTRAINT".to_string(),
+                    description: "cannot serve required skill".to_string(),
+                    detail: None
                 }]
             }]),
             ..create_empty_solution()